+++ pr_test/main.h
@@ -41,7 +41,7 @@ int print_uint(va_list, flags_t *);
 int print_hex(va_list, flags_t *);
 int print_octal(va_list, flags_t *);
 int print_HEX(va_list, flags_t *);
-int convert(unsigned int, int, int);
+int convert(unsigned long int, int, int);
 int num_of_digits(unsigned int);
 int print_bin(va_list, flags_t *);
 int get_flags(char, flags_t *);
